Time: 0 ms (100%), Space: 7.7 MB (68.61%) - LeetHub

+++ 0319-bulb-switcher/0319-bulb-switcher.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
     int bulbSwitch(int n) {
-        return static_cast<int>(sqrt(n));
+        return sqrt(n);
     }
 };
